+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add a new `plot-to-board` option to the burndown command to send the plotted
   burndown chart to the first card of the `Done` column.
+* Add documentation for all CLI commands. Fixes #83.
 
 ## Version 0.0.12